backend/requirement.txt
argon==0.2.0
bearer==3.1.0
distlib==0.3.1
Django==4.1.1
django-cleanup==6.0.0
django-cors-headers==3.10.1
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-restframework==0.0.1
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Faker==8.4.0
mysqlclient
Pillow==8.4.0
PyJWT==2.5.0
simplejwt==2.0.1
